requirements.txt
requests
python-dotenv
selenium
webdriver_manager
beautifulsoup4
lxml
python-amazon-sp-api
pandas
eel